% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP.R
\name{LCVP}
\alias{LCVP}
\title{Standardize plant names according to the Leipzig Catalogue of Plants (LCVP)}
\usage{
LCVP(
  splist,
  synonyms = TRUE,
  infra_specific = FALSE,
  genus_search = FALSE,
  max.distance = 0,
  encoding = "UTF-8",
  family_tab = FALSE,
  order_tab = FALSE,
  genus_tab = FALSE,
  save = FALSE,
  visualize = FALSE,
  version = "1.1",
  max.cores = (detectCores() - 1),
  out_path = getwd()
)
}
\arguments{
\item{splist}{A character vector specifying the input taxon, each element
including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name and author name}

\item{synonyms}{Logical. If TRUE, the full list of names, including synonyms 
will be returned for each name, if FALSE, only the accepted name will be returned.}

\item{infra_specific}{Logical. If TRUE, infra-specific names (e.g. subspecies) will be returned
for each submitted name.}

\item{genus_search}{Logical. If TRUE, the function will
apply the fuzzy match algorithm also for the search of the genus name,
otherwise as default the search is applied only to the epithet, the
infraspecies and the name author.}

\item{max.distance}{is an integer value. It represents the maximum distance
(number of characters) allowed for a match when comparing the submitted name
with the closest name matches in the LCVP}

\item{encoding}{character vector. This value will allow
the user to set the specific codification of the strings}

\item{family_tab}{Logical. If TRUE, the function will
return the list of plant taxa names belonging to the same family name
submitted by the user}

\item{order_tab}{Logical. If TRUE, the function will return
the list of plant taxa names belonging to the same order name submitted by
the user}

\item{genus_tab}{Logical. If TRUE, the function will return
the list of plant taxa names belonging to the same genus name submitted by
the user}

\item{save}{Logical. If TRUE, the function will write the
output file as comma-separated format (.csv), saving it into the working
directory or in the directory already set through the 'out_path' option}

\item{visualize}{If TRUE the function will visualize the output search on
the 'Source Tab' of RStudio. This option has to be turned off (FALSE) if the
package is execute in a UNIX environment (from the command line) without
having a Graphical User Interface}

\item{version}{A character vector indicating the current version of the
package. A new version is under development
allowing the package to connect the web API that is under construction}

\item{max.cores}{integer value, indicating the number of CPU cores to be
used for the parallelization of the plant name search when a list of plant
taxa names is submitted. As default, the maximum number of CPU cores
available on the working machine menus one is set}

\item{out_path}{a character vector. Defines the path
where the output file has to be saved.}
}
\description{
Allow a taxonomic resolution of plant taxa names listed in the "Leipzig
Catalogue of Vascular Plants" (LCVP). 
Connects to the LCVP table and validates the
names of a vector of plant taxa, replacing synonyms by accepted names and
removing orthographical errors in plant names. 
The LCVP data package must be installed. It is available from 
https://github.com/idiv-biodiversity/LCVP.
}
\examples{

LCVP("Hibiscus vitifolius")
LCVP("Hibiscus abelmoschus var. betulifolius Mast.")
LCVP(c("Hibiscus abelmoschus var. betulifolius Mast.", 
     "Hibiscus abutiloides Willd.", 
      "Hibiscus aculeatus", "Hibiscus acuminatus"), max.cores = 1)

}
\references{
The Leipzig Catalogue of Plants (LCVP) - An improved taxonomic
reference list for all known vascular plants
}
\seealso{
https://idata.idiv.de/ddm/Data/ShowData/1806
}
\author{
Alessandro Gentile, Alexander Zizka
}
\keyword{R-package}
\keyword{nomenclature}
\keyword{plants}
\keyword{taxonomy}
\keyword{vascular}
