% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP2.R
\name{LCVP2}
\alias{LCVP2}
\title{Standardize plant names according to the Leipzig Catalogue of Plants (LCVP)}
\usage{
LCVP2(splist)
}
\arguments{
\item{splist}{A character vector specifying the input taxon, each element
including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name and author name}
}
\description{
This is the new version of the LCVP function
}
