% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_group_search.R
\name{lcvp_group_search}
\alias{lcvp_group_search}
\title{Search plants by group in the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_group_search(
  group_names,
  search_by,
  max.distance = 0.1,
  bind_result = TRUE
)
}
\arguments{
\item{group_names}{A character vector specifying the group names. 
This includes names of order, family, genus or author.}

\item{search_by}{A character indicating whether to search by order ("Order"), 
family ("Family"), genus ("Genus") or authors ("Author").}

\item{max.distance}{It represents the maximum distance allowed for a match 
when comparing the submitted name with the closest name matches in the LCVP. 
Expressed either as integer, or as a fraction of the pattern length times the maximal 
transformation cost (will be replaced by the smallest integer not less than 
the corresponding fraction). See \code{\link[base]{agrep}} for more details.}

\item{bind_result}{If TRUE the function will return one data.frame. 
If False, the function will return a list of separate data.frames for 
each input group.}
}
\description{
Search plants by order, family, genus or author in the
Leipzig Catalogue of Plants (LCVP)
}
\examples{
\dontrun{
# By Genus 
res <- lcvp_group_search(c("AA", "Adansonia"), search_by = "Genus")
# By Author
res <- lcvp_group_search("Schltr.", search_by = "Author")

}
}
