% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_match.R
\name{lcvp_match}
\alias{lcvp_match}
\title{Match two name list using the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_match(
  splist1,
  splist2,
  max.distance = 0.1,
  include_all = TRUE,
  identify_dups = TRUE
)
}
\arguments{
\item{splist1}{A character vector specifying the reference input taxon to be matched.
Each element including genus and specific epithet and, potentially,
infraspecific rank, infraspecific name, and author name.}

\item{splist2}{A character vector specifying the input taxon to match splist1.
Each element including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name, and author name.}

\item{max.distance}{It represents the maximum distance allowed for a match
when comparing the submitted name with the closest name matches in the LCVP.
Expressed either as integer, or as a fraction of the pattern length times the maximal
transformation cost (will be replaced by the smallest integer not less than
the corresponding fraction). See \code{\link[base]{agrep}} for more details.}

\item{include_all}{If TRUE (default), it will include all species in both
splist1 and splist2. If FALSE, it will exclude species only found in splist2.}

\item{identify_dups}{If TRUE (default), a column indicating the position 
of duplicated LCVP output names in the resulting data.frame.}
}
\description{
Match two name list using the Leipzig Catalogue of Plants (LCVP)
}
\examples{
\dontrun{
splist1 <- sample(LCVP::tab_lcvp$Input.Taxon[5:100])
splist2 <- sample(LCVP::tab_lcvp$Input.Taxon[100:2])
res <- lcvp_match(splist1, splist2, include_all = TRUE)
}
}
