% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP.R
\name{lcvp_search}
\alias{lcvp_search}
\title{Standardize plant names according to the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_search(splist, max.distance = 0.1)
}
\arguments{
\item{splist}{A character vector specifying the input taxon, each element
including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name and author name}

\item{max.distance}{is an integer value. It represents the maximum distance
(number of characters) allowed for a match when comparing the submitted name
with the closest name matches in the LCVP}
}
\description{
This is the new version of the LCVP function
}
\examples{
\dontrun{

res_ex <- lcvp_search("Hibiscus vitifolius")
res_ex <- lcvp_search("Hibiscus vitifoliuse")
res_ex <- lcvp_search("Hibiscusa vitifolius")

res_ex <- lcvp_search("Hibiscus abelmoschus var. betulifolius Mast.")

res_ex <- lcvp_search(c("Hibiscus abelmoschus var. betulifolius Mast.",
"Hibiscus abutiloides Willd.",
 "Hibiscus aculeatus",
 "Hibiscus acuminatus"))

system.time(
res_ex <- lcvp_search(rep("Hibiscusa vitifolius", 10))
)
}
}
