% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_summary.R
\name{lcvp_summary}
\alias{lcvp_summary}
\title{Summary results for lcvp_search}
\usage{
lcvp_summary(lcvp_match)
}
\arguments{
\item{lcvp_match}{The output data.frame from lcvp_search.}
}
\description{
The functions summarizes the matching results from lcvp_search
}
\examples{
\dontrun{
res_ex <- lcvp_search(c("Hibiscus abelmoschus var. betulifolius Mast.",
"Hibiscus abutiloides Willd.",
"Hibiscus aculeatus",
"Hibiscus acuminatus",
"Hibiscus furcatuis",
"Hibiscus error"), 
max.distance = 1)
lcvp_summary(res_ex)
}
}
