% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvplants-defunct.R
\name{lcvplants-defunct}
\alias{lcvplants-defunct}
\alias{LCVP}
\alias{read_data}
\title{Defunct functions in package lcvplants}
\description{
These functions are defunct and no longer available: \code{LCVP} and
\code{read_data}.
}
\details{
These functions can be found in the older release v1.1.1 on the github page:
https://github.com/idiv-biodiversity/lcvplants/releases/tag/v1.1.1
}
