% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvplants-package.R
\docType{package}
\name{lcvplants-package}
\alias{lcvplants-package}
\alias{lcvplants}
\title{\Sexpr{tools:::Rd_package_title("lcvplants")}}
\description{
A package for large-scale taxonomic harmonization of plant names by fuzzy
matching and synonymy resolution against the Leipzig Plant Catalogue as
taxonomic backbone. Submission of single names or list of species names
is possible. The Leipzig Plant Catalogue is an updated taxonomic backbone based on the
plant list comprising more than 1,300,000 names and 350,000 accepted names
(this data is in the \href{https://github.com/idiv-biodiversity/LCVP}{\code{LCVP}}
package).
}
\section{lcvplants functions}{

\itemize{
\item \code{\link[lcvplants:lcvp_search]{lcvp_search}}: Standardize plant names
according to the LCVP
\item \code{\link[lcvplants:lcvp_summary]{lcvp_summary}}: Summarize the results
from lcvp_search
\item \code{\link[lcvplants:lcvp_group_search]{lcvp_group_search}}: Search
plants by taxa or author names in the LCVP
\item \code{\link[lcvplants:lcvp_fuzzy_search]{lcvp_fuzzy_search}}: Fuzzy match
plant names according to the LCVP
\item \code{\link[lcvplants:lcvp_join]{lcvp_join}}: Join two data.frames using the
LCVP
\item \code{\link[lcvplants:lcvp_match]{lcvp_match}}: Match two name lists using
the LCVP
\item \code{\link[lcvplants:lcvp_solve_dups]{lcvp_solve_dups}}: Solve duplicated
names by summarizing traits
}
}

\references{
The Leipzig Catalogue of Vascular Plants (LCVP) - An improved
taxonomic reference list for all known vascular plants.
}
\seealso{
https://idata.idiv.de/ddm/Data/ShowData/1806
}
\author{
Bruno Vilela
Alessandro Gentile
Martin Freiberg
Marten Winter
Alexander Zizka

Maintainer: Bruno Vilela \href{mailto:bvilela.bv@gmail.com}{bvilela.bv@gmail.com}
}
\keyword{R-package}
\keyword{nomenclature,}
\keyword{plants,}
\keyword{taxonomy,}
\keyword{vascular}
