% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Standardize plant names according to the Leipzig Catalogue of Plants (LCP)}
\usage{
read_data(pathstring, encoding)
}
\arguments{
\item{pathstring}{A character vector, which allow the user to define the
path where the input text file where is saved the input list of taxa is
stored.}

\item{encoding}{character vector, "UTF-8" (default). This value will allow
the user to set the specific codification of the strings.}
}
\description{
Allow a taxonomic resolution of plant taxa names listed in the "Leipzig
Catalogue of Plants" (LCP)
}
\examples{

\dontrun{read_data(file.choose(), 'UTF-8')}
}
\references{
The Leipzig Catalogue of Plants (LCP) - An improved taxonomic
reference list for all known vascular plants
}
\seealso{
https://idata.idiv.de/ddm/Data/ShowData/1806
}
\author{
Alessandro Gentile, Marten Winter, Martin Freiberg
}
