% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_fuzzy_search.R
\name{lcvp_fuzzy_search}
\alias{lcvp_fuzzy_search}
\title{Fuzzy match plant names according to the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_fuzzy_search(
  splist,
  max_distance = 0.2,
  genus_fuzzy = FALSE,
  status = c("accepted", "synonym", "unresolved", "external"),
  bind_result = TRUE,
  keep_closest = TRUE,
  progress_bar = FALSE
)
}
\arguments{
\item{splist}{A character vector specifying the input taxon, each element
including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name and author name. Only valid characters are allowed
(see \code{\link[base:validEnc]{validEnc}}).}

\item{max_distance}{It represents the maximum string distance allowed for a
match when comparing the submitted name with the closest name matches in the
LCVP. The distance used is a generalized Levenshtein distance that indicates
the total number of insertions, deletions, and substitutions allowed to
match the two names. It can be expressed as an integer or as the fraction of
the binomial name. For example, a name with length 10, and a max_distance =
0.1, allow only one change (insertion, deletion, or substitution). A
max_distance = 2, allows two changes.}

\item{genus_fuzzy}{If TRUE, the fuzzy match algorithm based on max_distance
will also be applied to the genus (note that this may considerably increase
computational time). If FALSE, fuzzy match will only apply to the epithet.}

\item{status}{A character vector indicating what taxa status should be
included in the results: "accepted", "synonym", "unresolved", "external".

The "unresolved" rank means that the status of the plant name could be
either valid or synonym, but the information available does not allow
a definitive decision. "external" is an extra rank that lists names
outside the scope of this publication but useful to keep on this
updated list.}

\item{bind_result}{If TRUE the function will return one data.frame (default).
If False, the function will return a list of separate data.frames for
each input group.}

\item{keep_closest}{if TRUE the function will return only the closest names
within the max_distance specified. If FALSE, it will return all names
within the specified distance.}

\item{progress_bar}{If TRUE, a progress bar will be printed.}
}
\value{
A data.frame or a list of data.frames (if \code{bind_result = FALSE})
with the following columns:

\itemize{
\item{\emph{Search}}{: Taxa name list provided by the user.}
\item{\emph{Input.Taxon}}{: The matched taxa name listed in the LCVP data.}
\item{\emph{Status}}{: Nomenclature status in the Leipzig Catalogue of
Plants: 'accepted', 'synonym', 'unresolved' or 'external'.
The "unresolved" rank means that the status of the plant name could be
either valid or synonym, but the information available does not allow
a definitive decision. "external" is an extra rank that lists names
outside the scope of this publication but useful to keep on this
updated list.}
\item{\emph{PL.comparisson}}{: This field provides a direct comparison with
‘The Plant List’ (TPL; The Plant List http://www.theplantlist.org/ accessed:
1.1. 2013) reporting further information such as ‘identical', 'synonym',
'other synonym', 'different authors', 'missing', 'misspelling',
'unresolved'.}
\item{\emph{PL.alternative}}{: This field provides a possible alternative
name from ‘The Plant List’.}
\item{\emph{Output.Taxon}}{: The list of the accepted plant taxa names
according to the LCVP. If input is 'accepted', the output is the same as the
input taxon.}
\item{\emph{Family}}{: The corresponding family name of the Input.Taxon,
staying empty if the Status is unresolved.}
\item{\emph{Order}}{: The corresponding order name of the Input.Taxon,
staying empty if the Status is unresolved.}
\item{\emph{Name.Distance}}{The approximate string distance between the Search
and matched Input.Taxon names. See \code{\link[utils:adist]{utils:adist}}
for more details.}
}
See \code{\link[LCVP:tab_lcvp]{LCVP::tab_lcvp}} for more details.

If no match is found for one species it will return NA for the columns in
the LCVP table. But, if no match is found for all species the function will
return NULL and a warning message.
}
\description{
Same as \code{\link[lcvplants:lcvp_search]{lcvp_search}}, but it returns
all matches from a fuzzy search of plant taxa names listed in the "Leipzig
Catalogue of Vascular Plants" (LCVP).
}
\details{
The algorithm will look for all the names within the given maximum distance
defined in \code{max_distance}. It can return all best matches (keep_closest =
TRUE), or all the matches within the distance (keep_closest = FALSE).

Note that only binomial names with valid characters are allowed in this
function. Search based on  genus, family, order or author names should use
the function  \code{\link[lcvplants:lcvp_group_search]{lcvp_group_search}}.
}
\examples{
# Ensure that LCVP package is available before running the example.
# If it is not, see the `lcvplants` package vignette for details
# on installing the required data package.
if (requireNamespace("LCVP", quietly = TRUE)) { # Do not run this

# Returns a data.frame
lcvp_fuzzy_search(c("Hibiscus vitifolia", "Artemisia vulgaris"))

# Returns a list of data.frames
lcvp_fuzzy_search(c("Hibiscus vitifolia", "Artemisia vulgaris"),
bind_result = FALSE)

# Returns all accepted names within a max_distance of 6.
lcvp_fuzzy_search("Hibiscus vitifolia", status = "accepted",
keep_closest = FALSE, max_distance = 6)

} 
}
\references{
Freiberg, M., Winter, M., Gentile, A. et al. LCVP, The Leipzig
catalogue of vascular plants, a new taxonomic reference list for all known
vascular plants. Sci Data 7, 416 (2020).
https://doi.org/10.1038/s41597-020-00702-z
}
\seealso{
\code{\link[lcvplants:lcvp_search]{lcvp_search}},
\code{\link[lcvplants:lcvp_group_search]{lcvp_group_search}}.
}
\author{
Bruno Vilela & Alexander Ziska
}
\keyword{R-package}
\keyword{nomenclature}
\keyword{plants}
\keyword{taxonomy}
\keyword{vascular}
