% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_group_search.R
\name{lcvp_group_search}
\alias{lcvp_group_search}
\title{Search plants by taxa or author names in the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_group_search(
  group_names,
  search_by,
  max.distance = 0.2,
  bind_result = TRUE,
  status = c("accepted", "synonym", "unresolved", "external")
)
}
\arguments{
\item{group_names}{A character vector specifying the taxa or author names.
This includes names of orders, families, genus or authors. Only valid
characters are allowed (see \code{\link[base:validEnc]{validEnc}}).}

\item{search_by}{A character indicating whether to search by "Order",
"Family", "Genus" or "Author".}

\item{max.distance}{It represents the maximum string distance allowed for a
match when comparing the submitted name with the closest name matches in the
LCVP. The distance used is a generalized Levenshtein distance that indicates
the total number of insertions, deletions, and substitutions allowed to
match the two names. It can be expressed as an integer or as the fraction of
the binomial name. For example, a name with length 10, and a max.distance =
0.1, allow only one change (insertion, deletion, or substitution). A
max.distance = 2, allows two changes.}

\item{bind_result}{If TRUE the function will return one data.frame (default).
If False, the function will return a list of separate data.frames for
each input group.}

\item{status}{A character vector indicating what taxa status should be
included in the results: "accepted", "synonym", "unresolved", "external".

The "unresolved" rank means that the status of the plant name could be
either valid or synonym, but the information available does not allow
a definitive decision. "external" is an extra rank that lists names
outside the scope of this publication but useful to keep on this
updated list.}
}
\value{
A data.frame or a list of data.frames (if \code{bind_result = FALSE})
with the following columns for all species of the matched groups:

\itemize{
\item{\emph{Input.Taxon}}{: Matched taxa names listed in the LCVP data.}
\item{\emph{Status}}{: Nomenclature status: 'accepted', 'synonym',
'unresolved' or 'external'.}
\item{\emph{PL.comparisson}}{: This field provides a direct comparison with
‘The Plant List’: ‘identical', 'synonym', 'other synonym',
'different authors', 'missing', 'misspelling' or 'unresolved'.}
\item{\emph{PL.alternative}}{: This field provides a possible alternative
name from ‘The Plant List’.}
\item{\emph{Output.Taxon}}{: The list of the accepted plant taxa names
according to the LCVP.}
\item{\emph{Family}}{: The corresponding family name of the Input.Taxon,
staying empty if the Status is unresolved.}
\item{\emph{Order}}{: The corresponding order name of the Input.Taxon,
staying empty if the Status is unresolved.}
}
See \code{\link[LCVP:tab_lcvp]{LCVP:tab_lcvp}} for more details.

If no match is found for all searched names the function will
return NULL and a warning message.
}
\description{
Search all plant taxa names listed in the "Leipzig
Catalogue of Vascular Plants" (LCVP) by order, family,
genus or author.
}
\details{
The algorithm will look for all the plant taxa names listed in the "Leipzig
Catalogue of Vascular Plants" (LCVP) based on a user-specified list of
orders,  families, genus, or authors names. If no match is found, it will try
to find the closest name given the maximum distance defined in \code{max.distance}.
If more than one name is fuzzy matched, only the first will be returned.
}
\examples{
# Ensure that LCVP package is available before running the example.
# If it is not, see the `lcvplants` package vignette for details
# on installing the required data package.
if (requireNamespace("LCVP", quietly = TRUE)) { # Do not run this
 
# Search by Genus
lcvp_group_search(c("AA", "Adansonia"), search_by = "Genus")

# Search by Author and keep only accepted names
lcvp_group_search("Schltr.", search_by = "Author", status = "accepted")

}
}
\references{
Freiberg, M., Winter, M., Gentile, A. et al. LCVP, The Leipzig
catalogue of vascular plants, a new taxonomic reference list for all known
vascular plants. Sci Data 7, 416 (2020).
https://doi.org/10.1038/s41597-020-00702-z
}
\seealso{
\code{\link[lcvplants:lcvp_search]{lcvp_search}},
\code{\link[lcvplants:lcvp_fuzzy_search]{lcvp_fuzzy_search}}.
}
\author{
Bruno Vilela & Alexander Ziska
}
\keyword{R-package}
\keyword{nomenclature}
\keyword{plants}
\keyword{taxonomy}
\keyword{vascular}
