% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_join.R
\name{lcvp_join}
\alias{lcvp_join}
\title{Join two data.frames using the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_join(
  x,
  y,
  sp_columns,
  max.distance = 0.1,
  type = "full",
  solve_duplicated = FALSE,
  func_numeric = mean,
  func_character = .keep_all,
  func_logical = any
)
}
\arguments{
\item{x, y}{data.frames to join.}

\item{sp_columns}{A character vector indicating the column names in x and y
with respective species names to join by.

For example, c("species", "Species_name").}

\item{max.distance}{It represents the maximum distance allowed for a match
when comparing the submitted name with the closest name matches in the LCVP.
Expressed either as integer, or as a fraction of the pattern length times the
maximal transformation cost (will be replaced by the smallest integer not
less than the corresponding fraction).
See \code{\link[base:agrep]{agrep}} for more details.}

\item{type}{What type of join should be done: "full" (default), "left", "right"
or "inner".
\itemize{
\item "full" return all rows and all columns from both x and y.
\item "left" return all rows from x.
\item "right" return all rows from y.
\item "inner" return all rows from x where there are matching species in y.
}}

\item{solve_duplicated}{if TRUE, it will summarize duplicated output names
given a function for each column class.

See \code{\link[lcvplants:lcvp_solve_dups]{lcvp_solve_dups}} for details.}

\item{func_numeric}{A function to summarize numeric columns
if solve_duplicated = TRUE. Default will return the mean.}

\item{func_character}{A function to summarize character columns
if solve_duplicated = TRUE. Default will keep all unique strings separated
by comma.}

\item{func_logical}{A function to summarize logical columns
if solve_duplicated = TRUE. Default will return TRUE if any is TRUE.}
}
\value{
A data.frame with the columns in both tables.
The rows will depend on the \code{type} selected. For "inner", a
subset of x rows. For "left", all x rows. For "right", a subset of x rows,
followed by unmatched y rows. For "full", all x rows, followed by unmatched
y rows.
}
\description{
Join two data.frames based on the taxonomic resolution of plant taxa names
listed in the "Leipzig Catalogue of Vascular Plants" (LCVP). Inspired by the
\code{\link[dplyr:join]{dplyr:join}} function.
}
\details{
The function add the columns from y to x
based on the list of species name in both tables. It
first standardizes the species names in both tables based on the "Leipzig
Catalogue of Vascular Plants" (LCVP) using the algorithm in
\code{\link[lcvplants:lcvp_search]{lcvp_search}}. These
standardized names of both tables are then matched using the algorithm in
\code{\link[lcvplants:lcvp_match]{lcvp_match}}. The type "full" join will
keep all species and add NAs to missing values. No NA is added in "inner",
"left" and "rigth" options.

Duplicated taxonomic resolution may occur if two inputs are now synonyms.
If \code{solve_duplicated} is \code{TRUE} the
\code{\link[lcvplants:lcvp_match]{lcvp_solve_dups}} function is applied to
merge duplicated output names.
}
\examples{
# Create data.frame1
splist1 <- sample(LCVP::tab_lcvp$Input.Taxon[2:100])
x <- data.frame("Species" = splist1, "Trait1" = runif(length(splist1)))

# Create data.frame2
splist2 <- sample(LCVP::tab_lcvp$Input.Taxon[98:8])
y <- data.frame("Species" = splist2,
"Trait2" = runif(length(splist2)),
"Trait3" = runif(length(splist2)),
"Trait4" = sample(c("a", "b"), length(splist2), replace = TRUE),
"Trait5" = sample(c(TRUE, FALSE), length(splist2), replace = TRUE))

# Full join
lcvp_join(x, y, c("Species", "Species"), type = "full")

# Left join
lcvp_join(x, y, c("Species", "Species"), type = "left")

# Right join
lcvp_join(x, y, c("Species", "Species"), type = "right")

# Inner join and solve duplicates
lcvp_join(x, y, c("Species", "Species"), 
type = "inner", solve_duplicated = TRUE)


}
\references{
Freiberg, M., Winter, M., Gentile, A. et al. LCVP, The Leipzig
catalogue of vascular plants, a new taxonomic reference list for all known
vascular plants. Sci Data 7, 416 (2020).
https://doi.org/10.1038/s41597-020-00702-z
}
\seealso{
\code{\link[lcvplants:lcvp_search]{lcvp_search}},
\code{\link[lcvplants:lcvp_match]{lcvp_match}},
\code{\link[lcvplants:lcvp_solve_dups]{lcvp_solve_dups}}.
}
\author{
Bruno Vilela & Alexander Ziska
}
\keyword{R-package}
\keyword{nomenclature}
\keyword{plants}
\keyword{taxonomy}
\keyword{vascular}
