% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_match.R
\name{lcvp_match}
\alias{lcvp_match}
\title{Match two name list using the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_match(splist1, splist2, max.distance = 0.1)
}
\arguments{
\item{splist1}{A character vector specifying the input taxon, each element
including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name and author name}

\item{splist2}{A character vector specifying the input taxon, each element
including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name and author name}

\item{max.distance}{It represents the maximum distance allowed for a match 
when comparing the submitted name with the closest name matches in the LCVP. 
Expressed either as integer, or as a fraction of the pattern length times the maximal 
transformation cost (will be replaced by the smallest integer not less than 
the corresponding fraction). See \code{\link[base]{agrep}} for more details.}
}
\description{
Match two name list using the Leipzig Catalogue of Plants (LCVP)
}
