% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_match.R
\name{lcvp_match}
\alias{lcvp_match}
\title{Match two name lists using the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_match(
  splist1,
  splist2,
  max.distance = 0.1,
  include_all = TRUE,
  identify_dups = TRUE
)
}
\arguments{
\item{splist1}{A character vector specifying the reference input taxon to be matched.
Each element including genus and specific epithet and, potentially,
infraspecific rank, infraspecific name, and author name. Only valid characters are allowed
(see \code{\link[base:validEnc]{base:validEnc}}).}

\item{splist2}{A character vector specifying the input taxon to match splist1.
Each element including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name, and author name. Only valid characters are allowed
(see \code{\link[base:validEnc]{base:validEnc}}).}

\item{max.distance}{It represents the maximum distance allowed for a match
when comparing the submitted name with the closest name matches in the LCVP.
Expressed either as integer, or as a fraction of the pattern length times the
maximal transformation cost (will be replaced by the smallest integer not
less than the corresponding fraction).
See \code{\link[base:agrep]{agrep}} for more details.}

\item{include_all}{If \code{TRUE} (default), it will include all species in both
\code{splist1} and \code{splist2}. If \code{FALSE}, it will exclude species
only found in \code{splist2}.}

\item{identify_dups}{If \code{TRUE} (default), a column indicating the position
of duplicated LCVP output names in the resulting data.frame.}
}
\value{
A data.frame with the following columns:

\itemize{
\item{\emph{Species.List.1}}{: Taxa name list provided by the user in the
splist1.}
\item{\emph{Species.List.2}}{: Taxa name list provided by the user in the
splist2.}
\item{\emph{LCVP.Input.Taxon}}{: Matched taxa names listed in the LCVP data.}
\item{\emph{Status}}{: Nomenclature status: 'accepted', 'synonym',
'unresolved' or 'external'.}
\item{\emph{LCVP.Output.Taxon}}{: The list of the accepted plant taxa names
according to the LCVP.}
\item{\emph{Match.Position.2to1}}{: positions of the names in splist1 in
splist2. Can be used to reorder splist2 to match splist1.
}
\item{\emph{Duplicated.Output.Position}}{: If \code{identify_dups = TRUE}, it
indicates the position of duplicated names in LCVP.Output.Taxon column.
This may occur if two inputs are now synonyms. It will output NA if there is
no duplicated for the species name.
}

}
See \code{\link[LCVP:tab_lcvp]{LCVP:tab_lcvp}} for more details.

If \code{include_all = TRUE}, all species will be included. Ordered based on the
\code{splist1}, and followed by non-matched names in \code{splist2}.
If \code{include_all = FALSE}, non-matched names in \code{splist2} are not
included.
}
\description{
Matches and compares two name lists based on the taxonomic resolution of
plant taxa names listed in the "Leipzig Catalogue of Vascular Plants" (LCVP).
}
\examples{

# Generate two lists of species name
splist1 <- sample(LCVP::tab_lcvp$Input.Taxon[5:100])
splist2 <- sample(LCVP::tab_lcvp$Input.Taxon[100:2])

# Including all species in both lists
lcvp_match(splist1, splist2, include_all = TRUE)

# Including all species only in the first list
matchLists <- lcvp_match(splist1, splist2, include_all = FALSE)
## This can be used to quickly change positions in splist2 to match splist1
splist2[matchLists$Match.Position.2to1]


}
\references{
Freiberg, M., Winter, M., Gentile, A. et al. LCVP, The Leipzig
catalogue of vascular plants, a new taxonomic reference list for all known
vascular plants. Sci Data 7, 416 (2020).
https://doi.org/10.1038/s41597-020-00702-z
}
\seealso{
\code{\link[lcvplants:lcvp_join]{lcvp_join}}
}
\author{
Bruno Vilela & Alexander Ziska
}
\keyword{R-package}
\keyword{nomenclature}
\keyword{plants}
\keyword{taxonomy}
\keyword{vascular}
