% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_search.R
\name{lcvp_search}
\alias{lcvp_search}
\title{Standardize plant names according to the Leipzig Catalogue of Plants (LCVP)}
\usage{
lcvp_search(
  splist,
  max.distance = 0.2,
  show.correct = FALSE,
  progress_bar = FALSE
)
}
\arguments{
\item{splist}{A character vector specifying the input taxon, each element
including genus and specific epithet and, potentially, infraspecific rank,
infraspecific name and author name. Only valid characters are allowed (see
\code{\link[base:validEnc]{base::validEnc}}).}

\item{max.distance}{It represents the maximum string distance allowed for a
match when comparing the submitted name with the closest name matches in the
LCVP. The distance used is a generalized Levenshtein distance that indicates
the total number of insertions, deletions, and substitutions allowed to
match the two names. It can be expressed as an integer or as the fraction of
the binomial name. For example, a name with length 10, and a max.distance =
0.1, allow only one change (insertion, deletion, or substitution). A
max.distance = 2, allows two changes.}

\item{show.correct}{If TRUE, a column is added to the final result indicating
whether the binomial name was exactly matched (TRUE), or if it is misspelled
(FALSE).}

\item{progress_bar}{If TRUE, a progress bar will be printed.}
}
\value{
A data.frame with the following columns: \itemize{
\item{\emph{Search}}{: Taxa name list provided by the user.}
\item{\emph{Input.Taxon}}{: The matched taxa name listed in the LCVP data.}
\item{\emph{Status}}{: Nomenclature status in the Leipzig Catalogue of
Plants: 'accepted', 'synonym', 'unresolved' or 'external'.
The "unresolved" rank means that the status of the plant name could be
either valid or synonym, but the information available does not allow
a definitive decision. "external" is an extra rank that lists names
outside the scope of this publication but useful to keep on this
updated list.}
\item{\emph{PL.comparisson}}{: This field provides a direct comparison with
‘The Plant List’ (TPL; The Plant List http://www.theplantlist.org/ accessed:
1.1. 2013) reporting further information such as ‘identical', 'synonym',
'other synonym', 'different authors', 'missing', 'misspelling',
'unresolved'.}
\item{\emph{PL.alternative}}{: This field provides a possible alternative
name from ‘The Plant List’.}
\item{\emph{Output.Taxon}}{: The list of the accepted plant taxa names
according to the LCVP. If input is 'accepted', the output is the same as the
input taxon.}
\item{\emph{Family}}{: The corresponding family name of the Input.Taxon,
staying empty if the Status is unresolved.}
\item{\emph{Order}}{: The corresponding order name of the Input.Taxon,
staying empty if the Status is unresolved.}
\item{\emph{Correct}}{: if show.correct = TRUE, this column is added to the
final result indicating whether the binomial name was exactly matched
(TRUE), or if it is misspelled (FALSE).} }

See \code{\link[LCVP:tab_lcvp]{LCVP::tab_lcvp}} for more
details.

If no match is found for one species it will return NA for the columns in the
LCVP table. But, if no match is found for all species the function will return
NULL and a warning message.
}
\description{
Allow taxonomic resolution of plant taxa names listed in the "Leipzig
Catalogue of Vascular Plants" (LCVP). Connects to the LCVP table and validates
the names of a vector of plant taxa, replacing synonyms with accepted names
and removing orthographic errors in plant names. The LCVP data package must be
installed. It is available from https://github.com/idiv-biodiversity/LCVP.
}
\details{
The function tries to match a name (Input.Taxon column) in LCVP, which has a
corresponding accepted valid name according to LCVP (Output.Taxon column). If
the Input.Taxon is a valid name, it will be the duplicated in Output.Taxon
column.

The algorithm will first try to exactly match the binomial names provided in
\code{splist}. If no match is found, it will try to find the closest name given the
maximum distance defined in \code{max.distance}. If more than one name is exactly
or fuzzy matched, only the accepted or the first will be returned and a
warning message will be printed on the console. The list of input names that
matched multiple names in LCVP can be obtained using \code{attr(x,
"matched_mult")}, being x the resulting data.frame. The function
\code{\link[lcvplants:lcvp_fuzzy_search]{lcvp_fuzzy_search}} can then be used
to return all results of the algorithm.

The \code{\link[lcvplants:lcvp_summary]{lcvp_summary}} function can be used to
summarize the results from a multiple species search, indicating the number of
species matched, and how many of them were exactly or fuzzy matched.

Note that only binomial names with valid characters are allowed in this
function. Search based on  genus, family, order or author names should use the
function \code{\link[lcvplants:lcvp_group_search]{lcvp_group_search}}.
}
\examples{
# Ensure that LCVP package is available before running the example.
# If it is not, see the `lcvplants` package vignette for details
# on installing the required data package.
if (requireNamespace("LCVP", quietly = TRUE)) { # Do not run this

# Search one species
lcvp_search("Aa achalensis")

# Search one species with misspelled name
lcvp_search("Aa achalensise", show.correct = TRUE)
lcvp_search("Aae achalensise", max.distance = 2)

# Search for a variety
lcvp_search("Hibiscus abelmoschus var. betulifolius Mast.")

# Search for multiple species
splist <- c(
"Hibiscus abelmoschus var. betulifolius Mast.",
"Hibiscus abutiloides Willd.",
"Hibiscus aculeatus",
"Hibiscus acuminatus",
"Hibiscus furcatuis" # This is a wrong name
)
mult <- lcvp_search(splist, max.distance = 0.2)

 ## Results for multiple species search can be summarized using lcvp_summary
lcvp_summary(mult)

}
}
\references{
Freiberg, M., Winter, M., Gentile, A. et al. LCVP, The Leipzig
catalogue of vascular plants, a new taxonomic reference list for all known
vascular plants. Sci Data 7, 416 (2020).
https://doi.org/10.1038/s41597-020-00702-z
}
\seealso{
\code{\link[lcvplants:lcvp_summary]{lcvp_summary}},
\code{\link[lcvplants:lcvp_group_search]{lcvp_group_search}},
\code{\link[lcvplants:lcvp_fuzzy_search]{lcvp_fuzzy_search}}.
}
\author{
Bruno Vilela & Alexander Ziska
}
\keyword{R-package}
\keyword{nomenclature}
\keyword{plants}
\keyword{taxonomy}
\keyword{vascular}
