% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvp_solve_dups.R
\name{lcvp_solve_dups}
\alias{lcvp_solve_dups}
\title{Solve duplicated names by summarizing traits}
\usage{
lcvp_solve_dups(
  x,
  duplicated_col,
  fixed_cols = NULL,
  func_numeric = mean,
  func_character = .keep_all,
  func_logical = any
)
}
\arguments{
\item{x}{data.frame.}

\item{duplicated_col}{The number of the column position with duplicated names
to be solved.}

\item{fixed_cols}{The columns positions that should be left out of the
summarizing processes. Normally applies for columns with fixed values across
repeated names.}

\item{func_numeric}{A function to summarize numeric columns
if solve_duplicated = TRUE. Default will return the mean.}

\item{func_character}{A function to summarize character or factor columns
if solve_duplicated = TRUE. Default will keep all unique strings separated
by comma.}

\item{func_logical}{A function to summarize logical columns
if solve_duplicated = TRUE.Default will return TRUE if any is TRUE.}
}
\value{
A data.frame with the same number of columns in  \code{x}
and combined duplicated lines according to functions provided.
}
\description{
Solve duplicated species names by summarizing traits given user provided
functions for common classes of variables (numeric, character, and logical).
}
\details{
The function will combine lines in \code{x} with duplicated names found in
\code{duplicated_col}. User-defined functions to combine the information in
\code{x} should take a vector (of length > 2) of the corresponding class
(numeric, character, and logical) and output only one value of the
corresponding class. Factors are transformed into characters.
}
\examples{

# Create a data.frame with duplicated names and different traits
splist <- sample(LCVP::tab_lcvp$Input.Taxon[1:100])
search <- lcvp_search(splist)

x <- data.frame("Species" = search$Output.Taxon,
"Trait1" = runif(length(splist)),
"Trait2" = sample(c("a", "b"), length(splist), replace = TRUE),
"Trait3" = sample(c(TRUE, FALSE), length(splist), replace = TRUE))

# Solve with default parameters
lcvp_solve_dups(x, 1)

# Summarize numbers using the median
lcvp_solve_dups(x, 1, func_numeric = median)

# Get one of characters at random
lcvp_solve_dups(x, 1, func_character = function(x){sample(x, 1)})


}
\author{
Bruno Vilela & Alexander Ziska
}
\keyword{R-package}
\keyword{nomenclature}
\keyword{plants}
\keyword{taxonomy}
\keyword{vascular}
